main.py

from pydantic import BaseModel, validator
from datetime import datetime, timedelta
from passlib.context import CryptContext
from enum import Enum
from typing import Union, Any, Annotated
import os
from jose import jwt
from bson import ObjectId
from schemas.task import taskEntity, tasksEntity
from schemas.user import userEntity, usersEntity
from config.db import db
from fastapi import FastAPI, HTTPException, Header
from routes.user import user
from routes.task import task

app = FastAPI()
app.include_router(user)
app.include_router(task)


ACCESS_TOKEN_EXPIRE_MINUTES = 30  # 30 minutes
REFRESH_TOKEN_EXPIRE_MINUTES = 60 * 24 * 7 # 7 days
ALGORITHM = "HS256"
JWT_SECRET_KEY = "secr"  # should be kept secret
JWT_REFRESH_SECRET_KEY = "secr2"    # should be kept secret

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def create_access_token(subject: Union[str, Any], expires_delta: int = None) -> str:
    if expires_delta is not None:
        expires_delta = datetime.utcnow() + expires_delta
    else:
        expires_delta = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode = {"exp": expires_delta, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, JWT_SECRET_KEY, ALGORITHM)
    return encoded_jwt

def create_refresh_token(subject: Union[str, Any], expires_delta: int = None) -> str:
    if expires_delta is not None:
        expires_delta = datetime.utcnow() + expires_delta
    else:
        expires_delta = datetime.utcnow() + timedelta(minutes=REFRESH_TOKEN_EXPIRE_MINUTES)
    
    to_encode = {"exp": expires_delta, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, JWT_REFRESH_SECRET_KEY, ALGORITHM)
    return encoded_jwt
