requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
motor
passlib[bcrypt]
passlib
python-jose[cryptography]
bson